requirements.txt
# Core ML
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.4
numpy==1.24.3

# MLflow for experiment tracking
mlflow==2.9.2

# Data validation
great-expectations==0.18.8

# API (for Phase 3)
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3

# Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

# Monitoring (for Phase 4)
prometheus-client==0.19.0

# Kubernetes client (for Phase 3)
kubernetes==28.1.0

# Docker SDK (for building images)
docker<7,>=4.0.0